improve ui

+++ TNBU.GUI/Pages/Configuration/WiFi/WiFiEdit.razor.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.EntityFrameworkCore;
 using MudBlazor;
+using TNBU.GUI.Components.Dialogs;
 using TNBU.GUI.EF;
 using TNBU.GUI.EF.Models;
 
@@ -9,11 +10,14 @@ namespace TNBU.GUI.Pages.Configuration.WiFi {
 	public partial class WiFiEdit : IDisposable {
 		[Inject] public IDbContextFactory<DB> DBS { get; set; } = null!;
 		[Inject] public NavigationManager NavigationManager { get; set; } = null!;
+		[Inject] public IDialogService DialogService { get; set; } = null!;
+		[Inject] public ISnackbar Snackbar { get; set; } = null!;
 
 		[Parameter] public long ID { get; set; }
 
 		private DB dbContext = null!;
 		private WiFiNetwork row = new();
+		private bool IsFormValid;
 
 		protected override void OnParametersSet() {
 			base.OnParametersSet();
@@ -32,13 +36,28 @@ namespace TNBU.GUI.Pages.Configuration.WiFi {
 				dbContext.WiFiNetworks.Add(row);
 			}
 			dbContext.SaveChanges();
+			Snackbar.Add($"WiFi network \"{row.SSID}\" saved", Severity.Success);
 			NavigationManager.NavigateTo("/configuration/wifi");
 		}
 
-		void OnDelete(MouseEventArgs e) {
-			dbContext.WiFiNetworks.Remove(row);
-			dbContext.SaveChanges();
-			NavigationManager.NavigateTo("/configuration/wifi");
+		async void OnDelete(MouseEventArgs e) {
+			var parameters = new DialogParameters {
+				{ "ContentText", $"Do you really want to delete \"{row.SSID}\"?" },
+			};
+
+			var options = new DialogOptions() {
+				CloseButton = true,
+				MaxWidth = MaxWidth.ExtraSmall,
+			};
+
+			var dialog = DialogService.Show<DeleteDialog>("Delete", parameters, options);
+			var result = await dialog.Result;
+			if(!result.Cancelled) {
+				dbContext.WiFiNetworks.Remove(row);
+				dbContext.SaveChanges();
+				Snackbar.Add($"WiFi network \"{row.SSID}\" deleted", Severity.Success);
+				NavigationManager.NavigateTo("/configuration/wifi");
+			}
 		}
 
 		public void Dispose() {
